Разделил FakePhotoRepository на FakePhotoCommandRepository и FakePhotoQueryRepository
